+++ ddRunSnippets_snippet.php
@@ -1,13 +1,13 @@
 <?php
 /**
  * ddRunSnippets
- * @version 4.0 (2023-03-30)
+ * @version 4.1 (2023-05-04)
  * 
  * @see README.md
  * 
- * @link https://code.divandesign.biz/modx/ddrunsnippets
+ * @link https://code.divandesign.ru/modx/ddrunsnippets
  * 
- * @copyright 2011–2023 DD Group {@link https://DivanDesign.biz }
+ * @copyright 2011–2023 Ronef {@link https://Ronef.ru }
  */
 
 //Include (MODX)EvolutionCMS.libraries.ddTools